feat: entry point optimizations
